requirements.txt
google-genai
rich
typer
